requirements.txt
discord
asyncio
srcomapi
requests
datetime
python-dotenv
gspread
discord-py-slash-command